requirements.txt
Django==3.0.6
django-material==1.6.3
django-viewflow==1.6.0
django-formtools==2.2
gunicorn==20.0.4
django-heroku